# apps/web/public/generated/openapi/openfoundry.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenFoundry API",
    "version": "0.1.0",
    "description": "Versioned OpenFoundry JSON/HTTP contract generated from proto services and curated REST overlays."
  },
  "servers": [
    {
      "url": "/",
      "description": "OpenFoundry API gateway root"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/pipelines/get-pipeline": {
      "get": {
        "summary": "PipelineService GetPipeline",
        "description": "Generated from `open_foundry.pipeline` RPC `GetPipeline` in service `PipelineService`.",
        "operationId": "open_foundry.pipeline.PipelineService.GetPipeline",
        "tags": [
          "open_foundry.pipeline"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "description": "Query parameter derived from `GetPipelineRequest.id`.",
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pipeline"
                }
              }
            }
          },
          "default": {
            "description": "Structured error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-openfoundry-sdk-namespace": "pipeline",
        "x-openfoundry-api-version": "v1",
        "x-openfoundry-mcp-tool": "openfoundry.pipeline.getpipeline",
        "x-openfoundry-stability": "beta"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {}
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ExternalConfig": {
        "type": "object",
        "properties": {
          "compute_profile_id": {
            "type": "string"
          },
          "source_id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "source_system": {
            "type": "string"
          }
        }
      },
...
      "IncrementalConfig": {
        "type": "object",
        "properties": {
          "allowed_transaction_types": {
            "type": "string"
          },
          "replay_on_deploy": {
            "type": "boolean"
          },
          "watermark_columns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Pipeline": {
        "type": "object",
        "properties": {
          "compute_profile_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "external": {
            "$ref": "#/components/schemas/ExternalConfig"
          },
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "incremental": {
            "$ref": "#/components/schemas/IncrementalConfig"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/PipelineLifecycle"
          },
          "name": {
            "type": "string"
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PipelineNode"
            }
          },
          "owner_id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "project_id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "status": {
            "type": "string"
          },
          "streaming": {
            "$ref": "#/components/schemas/StreamingConfig"
          },
          "type": {
            "$ref": "#/components/schemas/PipelineType"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PipelineNode": {
        "type": "object",
        "properties": {
          "config_json": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "id": {
            "type": "string"
          },
          "incremental_input": {
            "type": "boolean"
          },
          "input_dataset_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Uuid"
            }
          },
          "label": {
            "type": "string"
          },
          "output_dataset_id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "preview_status": {
            "type": "string"
          },
          "transform_type": {
            "type": "string"
          },
          "validation_errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "validation_status": {
            "type": "string"
          }
        }
      },
...
      "StreamingConfig": {
        "type": "object",
        "properties": {
          "input_stream_id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "output_stream_id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "parallelism": {
            "type": "integer",
            "format": "int32"
          },
          "streaming_profile_id": {
            "type": "string"
          }
        }
      },
...
      "Uuid": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token accepted by the OpenFoundry gateway. Session and personal access tokens share this scheme."
      }
    }
  }
}